backend/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
cryptography
pydantic
slowapi
